reset button resets both the nodes and Line highlighting
